requirements.txt
# Data Scraping
telethon==1.34.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# DBT
dbt-postgres==1.7.13

# Data Processing
pandas==2.2.0
numpy==1.24.3

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Orchestration
dagster==1.5.12
dagster-postgres==0.20.12

# Computer Vision
ultralytics==8.0.232
opencv-python-headless==4.9.0.80
pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2
tqdm==4.66.1

# API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0

# Computer Vision Dependencies
ultralytics==8.0.232
opencv-python-headless==4.9.0.80
pillow==10.1.0

# Data Processing
pandas==2.2.0
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
faker==24.11.0